refactor: rename dir, 파일 정리/#11
