+++ Application/Home/src/main/java/com/example/Home/Common/Repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.example.Home.MemberRepository;
+package com.example.Home.Common.Repository;
 
-import com.example.Home.Member.Member;
+import com.example.Home.Common.Entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
