--- Application/Home/src/main/java/com/example/Home/MemberRepository/MemberPointRepository.java
@@ -1,10 +1,0 @@
-package com.example.Home.MemberRepository;
-
-import com.example.Home.Member.MemberPoint;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberPointRepository extends JpaRepository<MemberPoint, Long> {
-    Optional<MemberPoint> findByMemberId(Long memberId);
-}
